closer-engine-data-layer/pyproject.toml
[project]
name = "closer-engine-data-layer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.7",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "pypdf>=6.7.5",
    "pytest-asyncio>=1.3.0",
    "supabase>=2.28.0",
]

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
